backend/pyproject.toml
[project]
name = "geo-llm-engine"
version = "0.1.0"
description = "MSCDP Capstone Project"
authors = [
    { name = "daegeun-kim", email = "kdgkimshsid@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pandas==2.3.3",
    "numpy==2.3.4",
    "geopandas==1.1.1",
    "scikit-learn==1.7.2",
    "openai>=2.8.0,<3.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "ipykernel>=7.1.0,<8.0.0",
    "sqlalchemy>=2.0.44,<3.0.0",
    "matplotlib>=3.10.7,<4.0.0",
    "psycopg2>=2.9.11,<3.0.0",
    "fastapi>=0.121.2,<0.122.0",
    "uvicorn>=0.38.0,<0.39.0",
    "lightgbm (>=4.6.0,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"